+++ data/scripts/galaxy/server.lua
@@ -8,14 +8,14 @@ function initialize(...)
         local galaxy = Galaxy()
         if galaxy then
             -- Attach background simulation scripts to the global galaxy loop
-            galaxy:addScriptOnce("data/scripts/server/background/cosmicwarnews.lua")
             galaxy:addScriptOnce("data/scripts/server/background/cosmicwardiplomaticsanctions.lua")
             galaxy:addScriptOnce("data/scripts/server/background/cosmicwarceasefires.lua")
             galaxy:addScriptOnce("data/scripts/server/background/cosmicwarbounties.lua")
             galaxy:addScriptOnce("data/scripts/server/background/cosmicwarbridgeupdate.lua")
             galaxy:addScriptOnce("data/scripts/server/background/cosmicwardiplomacy.lua")
-            galaxy:addScriptOnce("data/scripts/server/background/cosmicwarsiege_server.lua")
             galaxy:addScriptOnce("data/scripts/server/background/cosmicwarexpansion.lua")
+            galaxy:addScriptOnce("data/scripts/server/background/cosmicwarnews.lua")
+            galaxy:addScriptOnce("data/scripts/server/background/cosmicwarsiege_server.lua")
         end
     end
 end